Added operations Added 1 and 0 literals Started changing the query realisation

+++ src/main/scala/ru/hse/se/ba/danilin/paul/histogram/Histogram.scala
@@ -2,10 +2,10 @@ package ru.hse.se.ba.danilin.paul.histogram
 
 import ru.hse.se.ba.danilin.paul.histogram.atomizers.IAtomizer
 
-case class Histogram[O](histogram: Map[O, Int])(implicit universe: ElementsUniverse[O])
+case class Histogram[O](histogram: Map[O, Double])(implicit universe: ElementsUniverse[O])
   extends IHistogram[O] {
 
-  override def apply(element: O): Int = histogram.getOrElse(element, 0)
+  override def apply(element: O): Double = histogram.getOrElse(element, 0)
 
   override def elementsUniverse: ElementsUniverse[O] = universe
 
@@ -20,7 +20,7 @@ case class Histogram[O](histogram: Map[O, Int])(implicit universe: ElementsUnive
 object Histogram {
   def extract[S, O](source: S)(implicit atomizer: IAtomizer[S, O]): IHistogram[O] = {
     val atoms = atomizer atomize source
-    val histMap = atoms.foldLeft(Map.empty[O, Int])({
+    val histMap = atoms.foldLeft(Map.empty[O, Double])({
       case (acc, i) =>
         val newCount = acc.get(i) match {
           case None => 1
@@ -31,3 +31,5 @@ object Histogram {
     Histogram(histMap)
   }
 }
+
+
