+++ src/main/scala/ru/hse/se/ba/danilin/paul/histogram/operations/Subtract.scala
@@ -1,7 +1,7 @@
 package ru.hse.se.ba.danilin.paul.histogram.operations
 
 object Subtract extends HistogramBinaryMergeOperation {
-  override protected def merge(leftCount: Int, rightCount: Int): Int =
+  override protected def merge(leftCount: Double, rightCount: Double): Double =
     if (rightCount > 0)
       0
     else leftCount
