+++ src/main/scala/ru/hse/se/ba/danilin/paul/histogram/operations/Unite.scala
@@ -1,8 +1,8 @@
 package ru.hse.se.ba.danilin.paul.histogram.operations
 
 object Unite extends HistogramBinaryMergeOperation {
-  def max = (a: Int, b: Int) => if (a > b) a else b
+  def max = (a: Double, b: Double) => if (a > b) a else b
 
-  override protected def merge(leftCount: Int, rightCount: Int): Int =
+  override protected def merge(leftCount: Double, rightCount: Double): Double =
     max(leftCount, rightCount)
 }
